📝 feat. 회원탈퇴Modal  코드 정리
