+++ src/components/modal/withdrawalModal/WithdrawalModal.tsx
@@ -1,31 +1,20 @@
 import { useState, useEffect } from "react";
 import styles from "./withdrawalModal.module.scss";
 import { useNavigate } from "react-router-dom";
-//elements
 import { ModalCancelBtn, ModalDelBtn } from "../../../elements/buttons/Buttons";
-//hook
 import useModal from "../useModal";
-import { useWithdrawal } from "./useWithdrawal";
-//mui toast
+import { useWithdrawalQuery } from "./useWithdrawalQuery";
 import Snackbar from "@mui/material/Snackbar";
 import MuiAlert from "@mui/material/Alert";
+import { WithdrawalType } from "./withdrawalModal.type";
 
-export interface WithdrawalType {
-  title: string;
-}
-
-const WithdrawalModal = (props: WithdrawalType) => {
-  const { title } = props;
+const WithdrawalModal = ({ title }: WithdrawalType) => {
   const navigate = useNavigate();
-  //에러메세지
   const [errMsg, setErrMsg] = useState<string>();
   //toast alert on/off
   const [open, setOpen] = useState(false);
-  //모달 닫기
   const { hideModal } = useModal();
-  //회원탈퇴 hook
-  const { mutate, isError, isSuccess, error } = useWithdrawal();
-
+  const { mutate, isError, isSuccess, error } = useWithdrawalQuery();
   //회원탈퇴 성공시
   if (isSuccess) {
     localStorage.clear();
@@ -34,11 +23,9 @@ const WithdrawalModal = (props: WithdrawalType) => {
   }
   //회원탈퇴 에러처리
   useEffect(() => {
-    if (isError) {
-      console.log(error.response.data.errorMessage);
-      setErrMsg(error.response.data.errorMessage);
-      setOpen(true);
-    }
+    if (!isError) return;
+    setErrMsg(error.response.data.errorMessage);
+    setOpen(true);
   }, [isError]);
 
   return (
@@ -47,8 +34,7 @@ const WithdrawalModal = (props: WithdrawalType) => {
         <h2>{title}</h2>
         <div className={styles.btnBox}>
           <ModalCancelBtn onClick={() => hideModal()}>취소</ModalCancelBtn>
-          <ModalDelBtn>탈퇴하기</ModalDelBtn>
-          {/* <ModalDelBtn onClick={() => mutate()}>탈퇴하기</ModalDelBtn> */}
+          <ModalDelBtn onClick={() => mutate()}>탈퇴하기</ModalDelBtn>
         </div>
       </div>
       <Snackbar
