+++ src/components/modal/withdrawalModal/withdrawalModal.type.ts
@@ -1,0 +1,6 @@
+export interface WithdrawalType {
+  title: string;
+}
+export interface ErrType {
+  errorMessage: string;
+}
